remove axios from file
